Initial cut at gradle build logic
Gradle is a great build system. It's worth migrating to because it will
allow us to push new releases of jadretro into the maven central
repository easily. Running './gradle build' will create a jar in
build/lib/ that is identical to the jar built by ant. With a little
more effort, all build functionality in ant will be in gradle and ant
can be jettisoned.